Minor improvements (#73)
-Improved the warning message shown when importing unrecognised data. The displayed message is now less alarmist.
-Removed the unnecessary warning in the switches screen under system settings.
-Removed some more redundant variables in the source code.

+++ source code/transmitter/mtx/src/common.cpp
@@ -1115,8 +1115,8 @@ void trimWhiteSpace(char* buff, uint8_t lenBuff)
 void sanitize83BaseName(char *str)
 {
   //convert to upper case
-  size_t len = strlen(str);
-  for(size_t i = 0; i < len; i++)
+  uint8_t len = strlen(str);
+  for(uint8_t i = 0; i < len; i++)
     str[i] = toupper(str[i]);
   
   //replace invalid characters
